convert refresh token part to a separate service
